Move library to root of repository
Previously this repository had a strange structure of some library files (keywords.txt, examples) in the repository root but the library source files under library/abstractIO.

A popular library installation technique is to download the library via GitHub's Clone or download > Download ZIP and then use the Arduino IDE's Sketch > Include Library > Add .ZIP Library on the downloaded file. This requires the library to be in the root of the repository, not in a subfolder. If the library is not in the root of the repository this installation technique fails:

Specified folder/zip file does not contain a valid library

This move is also required if you wanted to add your library to the Arduino Library Manager index, which provides an even easier installation option.

Since it appears you prefer the library source files to be stored in a subfolder, I have chosen the Arduino Library 1.5 format (source files in the src subfolder). This does break backwards compatibility with Arduino IDE 1.0.x.